Create the volume component.
